requirements.txt
black>=25.9.0
mypy>=1.18.2
pylint>=3.3.8
pymongo>=4.15.2
pytest>=8.4.2
python-dotenv>=1.1.1
rich>=14.1.0
ruff>=0.13.3
